requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
Django==2.1.2
idna==2.7
mysqlclient==1.3.13
pycparser==2.19
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
